Rename funding.ts to directFunding.ts
