+++ packages/wallet/src/redux/states/channels/funding/directFunding.ts
